Fix : 마우스를 화면에 올렸을 때 ButtonBar의 작동 로직 변경
- 기존에 마우스가 화면 위에 있는지 확인하는 state가 ToggleStore에 존재하고, 그것들을 MainScreen이 구독하고 있기 때문에 리랜더링이 너무 자주 발생하는 부작용이 있었습니다.
- 따라서 isMouseOnScreen state를 ButtonBar 내부로 옮기고, 작동 로직을 변경하였습니다.
